pyproject.toml
[tool.poetry]
name = "legal_case_search"
version = "0.1.0"
description = ""
authors = ["Songjae Lee <mmqq2003@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
streamlit = "^1.40.2"
openai = "^1.56.1"
langchain = "^0.3.9"
langchain-openai = "^0.2.11"
python-dotenv = "^1.0.1"
pandas = "^2.2.3"
faiss-cpu = "^1.9.0.post1"
langchain-community = "^0.3.9"
unstructured = "^0.16.9"
openpyxl = "^3.1.5"
networkx = "^3.4.2"
kiwipiepy = "^0.20.3"
rank-bm25 = "^0.2.2"
tavily-python = "^0.5.0"



[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
